[hotfix] Harden TaskExecutorTest#testSyncSlotsWithJobMasterByHeartbeat
The problem was that we don't wait until the JobMaster has sent its response back
to the TaskExecutor when it receives the offered slots and before we send the
heartbeat response with the AllocatedSlotReport. This commit fixes the problem
by using the FutureUtils#whenCompleteAsyncIfNotDone to avoid the inseration of
another runnable for the test case.