Remove testFailingAllocationReturnsPendingTaskManagerSlot
Canceling slot allocations if the requirements go down is somewhat unnecessary since the slot will eventually either be used for other purposes or end up as a free slot  (if upon allocation no matching requirement exists).
We do risk that this slot sits around for a while (if another slot of the same TM is allocated), but we need some mechanism to compress slots into as few as desired task executors eventually anyway.