Move allocation calls out of requirements tracker
There is a weird inconsistency with the tracker issuing allocate calls immediately when processing requirements, or on demand when called from the outside.
We were also leaking references to the slot manager while it was under construction. Finally, the current design also makes things less flexible; the tracker would need a lot of additional information to make informed decisions about which should should get how many resources. It is better to relegate this logic to another component, and have to just be a tracker that can determine how many resources are exceeding/required.