Decouple resource tracker from slot life-cycle
The JobResourceState was unnecessarily mirroring the slot life-cycle, and the tracker as a whole was too tightly coupled with it. That a slot can be pending/allocated is an implementation detail of the allocation protocol, and the distinction between the two is irrelevant in terms of fulfilling resources; all we care about is that a slot has been reserved for a specific job. When this resource is actually fulfilled in practical terms is not relevant for the tracker.