Avoid notifying new slots multiple times when trying to allocate a shared slot
The problem is that the SchedulerImpl allocates and releases available slots when
trying to allocate a shared slot. Due to this new slots will be notified multiple
times to the FutureSlotPool since every release will call notifyNewSlotsAvailable
in the DeclarativeSlotPoolNg. As a consquence, it can happen that we try to
assign multiple times the same JobVertexId to one shared slot.