Fix race condition between fulfilling pending requests and newly created requests
If we don't fulfill first the pending requests with newly offered slots, it
can happen that we fulfill the first request, then creating new requests which
use other newly offered slots. The problem is that the new requests won't
use an unresolved shared slot if it can obtain an available slots. This results
in a scheduling where tasks which can share a slot end up in separate slots meaning
that the job needs more slots than necessary to run.