Early exit if no pending requests exist
There are tests that call close() after suspend(). In this case the jobmanager targetAddress is already null. Within close() the FutureSlotPool tries to clear all pending resources, and calls into the declarative slot pool regardless of whether there are requests to cancel or not. This triggers a requirement "reduction" by 0, for which a ResourceRequirements is created, but the constructor rejects null target addresses.